Only load a config file if it exists.
This also fixes the problem when installing the gem as a rails plugin, the load paths are hard coded into the plugin's init file and thus the plugin throws error.

+++ lib/app_config.rb
@@ -50,7 +50,7 @@ private
   end
   
   def load_conf_file(conf_path)
-    return {} if !conf_path or conf_path.empty?
+    return {} if !conf_path or conf_path.empty? or !File.exist?(conf_path)
     File.open(conf_path, "r") do |file|
       YAML.load(ERB.new(file.read).result) || {}
     end
